% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMMs.R
\name{check.missing}
\alias{check.missing}
\title{check.missing}
\usage{
check.missing(X, pathways, method, allgenes)
}
\arguments{
\item{X}{new data}

\item{pathways}{one of the "hall","biocarta","pid","cgp","wiki","tft.legency",
"tft.gtrd", "cm", "gobp", "gocc", "gomf", "oncogene", "immune", "vax" and "genes".
If pathways = "genes", the gene based model is selected.}

\item{method}{one of "ssgsea", "gsva", "zscore", "grp", if models are constructed
based on pathways.}

\item{allgenes}{all genes in the training data}
}
\description{
Check missing variables for prediction. The function checks the missing values of
the new data with respect to the training data, including the missing background
values and missing model values. Missing background values are not directly involved
in the fitted models, but they may contribute to the calculation of pathway scores,
e.g., gsva.
}
\examples{
NewX <- Xtrain[-c(1:100), 1:10]
check.missing(NewX, pathways = "vax", method = "grp", allgenes = rownames(Xtrain)) 

}
