% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMMs.R
\name{miss.impute}
\alias{miss.impute}
\title{miss.impute}
\usage{
miss.impute(NewX, Xtrain)
}
\arguments{
\item{NewX}{new data}

\item{training}{training data used for model building, GSE136324}
}
\description{
Imputed missing values using K-nearest neighbors based on the training data (GSE136324).
}
\examples{
NewX <- Xtrain[-c(1:100), 1:10]
miss.impute(NewX, Xtrain)

}
