% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMMs.R
\name{predictMM}
\alias{predictMM}
\title{predictMM}
\usage{
predictMM(
  X,
  pathways = "immune",
  method = "ssgsea",
  miss.impute = T,
  type = c("lp", "risk", "expected", "terms", "survival")
)
}
\arguments{
\item{X}{variables of new data}

\item{pathways}{one of 14 sources of pathways: "hall","biocarta","pid","cgp","wiki","tft.legency",
"tft.gtrd", "cm", "gobp", "gocc", "gomf", "oncogene", "immune","vax".
If pathways = "genes", a gene based model is selected.}

\item{method}{one of "ssgsea", "gsva", "zscore", "grp", if models are constructed
based on pathways.}

\item{miss.impute}{whether missing data should be imputed.}

\item{type}{the type of predicted value. Choices are the linear predictor ("lp"),
the risk score exp(lp) ("risk"), the expected number of events given the
covariates and follow-up time ("expected"), and the terms of the linear predictor
("terms"). The survival probability for a subject is equal to exp(-expected).
see predict.coxph in survival package.}
}
\description{
The function provided a set of prognostic models constructed either based on
genes or pathways. For the pathway models, two strategies were implemented to construct predictive
models: pathway score method and group lasso using pathway information.
We considered three methods for pathway score calculation (ssGSEA, GSVA, and z-scores) and
14 data sources providing pathway information.
We recommended a group lasso model based on the vax pathways.
}
\examples{
NewX <- Xtrain[-c(1:100), 1:10]
predictMM(NewX, pathways = "vax", method = "grp", type = "lp")

}
